Remove old code, fix imports and tests
