+++ src/bloqade/cirq_utils/noise/two_zone_utils.py
@@ -6,8 +6,6 @@ import cirq
 import numpy as np
 from cirq.circuits.qasm_output import QasmUGate
 
-from bloqade.qasm2.dialects.noise import TwoRowZoneModel
-
 Slot = Tuple[int, int]  # (tuple index, position inside tuple)
 Swap = Tuple[Slot, Slot]
 
@@ -531,91 +529,3 @@ def extract_u3_and_cz_qargs(moment: cirq.Moment):
             result["cz"].append(op.qubits)
 
     return result
-
-
-def get_two_zoned_noisy_circ(
-    circ: cirq.Circuit,
-    move_rates=None,
-    sq_loc_rates=None,
-    sq_glob_rates=None,
-    cz_rates=None,
-    sitter_rates=None,
-    unp_cz_rates=None,
-):
-    """
-    Applies noise to a Cirq circuit with two zones, where the noise is applied to both the gates and the movement of qubits.
-
-    Args:
-        circ: A cirq.Circuit object.
-        move_rates: parameters for move noise channel
-        sq_loc_rates: parameters for local qubit gate channel
-        cz_rates: parameters for two qubit gate noise channel
-    Returns:
-        A new cirq.Circuit object with noise applied.
-    """
-    if move_rates is None:
-        px_move = TwoRowZoneModel().mover_px
-        py_move = TwoRowZoneModel().mover_py
-        pz_move = TwoRowZoneModel().mover_pz
-        move_rates = [px_move, py_move, pz_move]  # default move error
-    if sq_loc_rates is None:
-        sq_loc_rates = TwoRowZoneModel().local_errors[
-            0:3
-        ]  # default local single qubit gate error rates
-    if sq_glob_rates is None:
-        sq_glob_rates = TwoRowZoneModel().global_errors[
-            0:3
-        ]  # default global single qubit gate error rates
-    if cz_rates is None:
-        cz_rates = TwoRowZoneModel().cz_paired_errors[0:3]
-        cz_rates = (
-            cz_rates + cz_rates
-        )  # control and target qubits decohere with same probabilities
-    if sitter_rates is None:
-        sit_px = TwoRowZoneModel().sitter_px
-        sit_py = TwoRowZoneModel().sitter_py
-        sit_pz = TwoRowZoneModel().sitter_pz
-        sitter_rates = [sit_px, sit_py, sit_pz]
-    if unp_cz_rates is None:
-        unp_cz_rates = TwoRowZoneModel().cz_unpaired_errors[0:3]
-
-    qbits = circ.all_qubits()
-    nqubs = len(qbits)
-    noisy_circ = cirq.Circuit()
-
-    are_all_named = all(isinstance(q, cirq.NamedQubit) for q in qbits)
-    if are_all_named:
-        mapping = get_map_named_to_line_qubits(qbits)  # type: ignore -- linter being dumb
-        circ = circ.transform_qubits(lambda q: mapping[q])
-
-    for i in range(len(circ)):
-
-        if i == 0:
-            noisy_circ.append(
-                get_move_error_channel_two_zoned(
-                    circ[i], None, np.array(move_rates), np.array(sitter_rates), nqubs
-                )
-            )
-        else:
-            noisy_circ.append(
-                get_move_error_channel_two_zoned(
-                    circ[i],
-                    circ[i - 1],
-                    np.array(move_rates),
-                    np.array(sitter_rates),
-                    nqubs,
-                )
-            )
-
-        noisy_circ.append(circ[i])
-        noisy_circ.append(
-            get_gate_error_channel(
-                circ[i],
-                np.array(sq_loc_rates),
-                np.array(sq_glob_rates),
-                np.array(cz_rates),
-                np.array(unp_cz_rates),
-            )
-        )
-
-    return noisy_circ
