--- test/cirq_utils/noise/test_two_zone_model.py
@@ -1,49 +1,0 @@
-import cirq
-
-from bloqade.cirq_utils.noise import (
-    GeminiTwoZoneNoiseModel,
-    transform_circuit,
-    get_two_zoned_noisy_circ,
-)
-
-
-def create_ghz_circuit(n):
-    qubits = cirq.LineQubit.range(n)
-    circuit = cirq.Circuit()
-
-    # Step 1: Hadamard on the first qubit
-    circuit.append(cirq.H(qubits[0]))
-
-    # Step 2: CNOT chain from qubit i to i+1
-    for i in range(n - 1):
-        circuit.append(cirq.CNOT(qubits[i], qubits[i + 1]))
-
-    return circuit
-
-
-model = GeminiTwoZoneNoiseModel()
-
-circuit = create_ghz_circuit(2)
-
-circuit_ = cirq.optimize_for_target_gateset(circuit, gateset=cirq.CZTargetGateset())
-print(circuit_)
-noisy_circuit = get_two_zoned_noisy_circ(circuit_)
-
-print(noisy_circuit)
-
-compressed_circuit_from_model, noisy_circuit_from_model = transform_circuit(
-    circuit, model=model
-)
-
-print(compressed_circuit_from_model)
-print(noisy_circuit_from_model)
-
-new_moments = []
-for idx, moment in enumerate(noisy_circuit):
-    if moment == cirq.Moment():
-        continue
-    else:
-        new_moments.append(moment)
-noisy_circuit2 = cirq.Circuit.from_moments(*new_moments)
-
-assert noisy_circuit2 == noisy_circuit_from_model
